changing name to invalid organisation and including test cases
