+++ digital_land/phase/organisation.py
@@ -18,9 +18,10 @@ class OrganisationPhase(Phase):
                 self.issues.line_number = block["line-number"]
                 self.issues.entry_number = block["entry-number"]
 
+            organisation_value = row["organisation"]
             row["organisation"] = self.organisation.lookup(row["organisation"])
             if not row["organisation"]:
                 self.issues.log_issue(
-                    "organisation", "missing organisation", row["organisation"]
+                    "organisation", "invalid organisation", organisation_value
                 )
             yield block
